Fix stale icon for gamecard titles on HOME Menu
Previously, Mikage would always return title ID 0 for any gamecard
image. This would make the HOME Menu assume that *the same* gamecart is
inserted every time even when a different one is used, causing the HOME
Menu to display the incorrect icon, as the HOME Menu stores title icons
in a cache entry that is populated only the first time the icon appears
on the HOME Menu.